GUI Sprites
Created the HUD elements for the player's health, as well as the folder to contain the GUI elements in general.
Use the first 3 sprites in the rows to simulate a heartbeat animation for the rightmost (currently active) heart if possible. When taking damage, use the heartbeat animation + the transitional sprite (which is rightmost on the spritesheet. The HUD element is what the hearts should be displayed over.